Operator Rotation Facing VFX
When the operator looks in certain directions, more VFX details will change based on where they look. it used to just be the sprite changing when looking front and back.

Now, the operator can wear stuff on their "front" or "back", and the left and right hand item positions are maintianed.

all of this is to say that despite the game being 2D, the way the equipment looks on the operator is realistic regardless of where they look. No randomly switching hands, wierd clipping.