Slow Aiming & Updated VFX to match aiming position
Aiming VFX is complete for single item and akimbo item stances. Aiming shouldn't be a concern from here on out.

items use a 0-1 scale to affect how quickly they can rotate to a target position. Rotation is done very smoothly, which took a lot of reworking the old VFX system